+++ BookApp.Core/Services/CategoryServices.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using BookApp.Core.DTO;
 using BookApp.Core.Interfaces;
@@ -15,7 +16,7 @@ namespace BookApp.Core.Services
 
         public CategoryServices(ICategoryRepository categoryRepository)
         {
-            _categoryRepository = categoryRepository;
+            _categoryRepository = categoryRepository ?? throw new ArgumentNullException(nameof(categoryRepository));
         }
 
         public async Task<List<CategoryDTO>> GetAllCategoriesAsync()
@@ -25,21 +26,41 @@ namespace BookApp.Core.Services
 
         public async Task<CategoryDTO> GetCategoryByIdAsync(int id)
         {
+            if (id <= 0)
+            {
+                throw new ArgumentException("Invalid category ID", nameof(id));
+            }
+
             return await _categoryRepository.GetCategoryByIdAsync(id);
         }
 
         public async Task AddCategoryAsync(CategoryDTO category)
         {
+            if (category == null)
+            {
+                throw new ArgumentNullException(nameof(category));
+            }
+
             await _categoryRepository.AddCategoryAsync(category);
         }
 
         public async Task UpdateCategoryAsync(CategoryDTO category)
         {
+            if (category == null)
+            {
+                throw new ArgumentNullException(nameof(category));
+            }
+
             await _categoryRepository.UpdateCategoryAsync(category);
         }
 
         public async Task DeleteCategoryAsync(int id)
         {
+            if (id <= 0)
+            {
+                throw new ArgumentException("Invalid category ID", nameof(id));
+            }
+
             await _categoryRepository.DeleteCategoryAsync(id);
         }
     }
